feat: Added/Updated benchmark problems (preliminary)
The commited code is still in a preliminary state and missing proper
documentation and test cases, so handle everything with care.

There is now a specialised (but still marked as abstract) class called
BechnmarkProblem to provide methods for reocurring calculations within
many benchmark problems. It also defines boundaries and the
shift/rotation of fitness landscapes according to the Real-Parameter
Black-Box Optimization Benchmark 2013 from INRIA.

Addition of the following benchmark problems, based on the
Real-Parameter Black-Box Optimization Benchmark 2013.
- Attrachtive Sector Function
- Bent Cigar Function
- Büche-Rastrigin Function
- Discus Function
- Ellipsoidal Function, rotated
- Step Ellipsoidal Function

Updated the following problems based on the new abstract class
BenchmarkProblem
- Ellipsoidal Function
- Linear Slope
- Rastrigin Function
- Sphere Function

Signed-off-by: Sebastian Niemann <niemann@sra.uni-hannover.de>